Safari-611.1.10/LayoutTests/fast/events/ios/fire-input-and-keypress-on-return-key-expected.txt

Tests that pressing the Return key in a content editable elemnent dispatches a DOM input event and DOM keypress event (in that order). To run this test by hand, focus the content editable element below and press the Return key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.key is "Enter"
PASS event.inputType is "insertParagraph"
PASS successfullyParsed is true

TEST COMPLETE
